+++ packages/narada/src/narada/window.py
@@ -6,9 +6,7 @@ from http import HTTPStatus
 from typing import Any, Generic, Literal, TypedDict, TypeVar, overload
 
 import aiohttp
-from playwright.async_api import BrowserContext
-from pydantic import BaseModel
-
+from narada.config import BrowserConfig
 from narada_core.actions.models import (
     AgenticSelectorAction,
     AgenticSelectorRequest,
@@ -23,25 +21,28 @@ from narada_core.actions.models import (
     ReadGoogleSheetResponse,
     WriteGoogleSheetRequest,
 )
-from narada_core.models import Agent, RemoteDispatchChatHistoryItem, UserResourceCredentials
 from narada_core.errors import NaradaError, NaradaTimeoutError
-from narada_core.responses import (
+from narada_core.models import (
+    Agent,
+    RemoteDispatchChatHistoryItem,
     Response,
     ResponseContent,
     Usage,
-    _StructuredOutput,
+    UserResourceCredentials,
     _MaybeStructuredOutput,
     _ResponseModel,
+    _StructuredOutput,
 )
 from narada_core.window import BaseBrowserWindow
-from narada.config import BrowserConfig
+from playwright.async_api import BrowserContext
+from pydantic import BaseModel
 
 # Note: Core types now imported from narada_core above
 
 
 class NaradaBrowserWindow(BaseBrowserWindow):
     """Browser window implementation for the narada package using aiohttp."""
-    
+
     _api_key: str
     _base_url: str
